feat: [PIPE-21479]: add pr commits page - fix utils
