+++ packages/playground/src/components/pull-request/utils.tsx
@@ -93,19 +93,3 @@ export const timeDistance = (date1 = 0, date2 = 0, onlyHighestDenomination = fal
     minutes ? minutes + 'm' : hours || days ? '0m' : ''
   } ${seconds}s`
 }
-const LOCALE = Intl.NumberFormat().resolvedOptions?.().locale || 'en-US'
-
-/**
- * Format a timestamp to medium format date (i.e: Jan 1, 2021)
- * @param timestamp Timestamp
- * @param dateStyle Optional DateTimeFormat's `dateStyle` option.
- */
-export function formatDate(timestamp: number | string, dateStyle = 'medium'): string {
-  return timestamp
-    ? new Intl.DateTimeFormat(LOCALE, {
-        // eslint-disable-next-line @typescript-eslint/ban-ts-comment
-        // @ts-ignore: TS built-in type for DateTimeFormat is not correct
-        dateStyle
-      }).format(new Date(timestamp))
-    : ''
-}
